+++ DCHdigi/include/DCHsimpleDigitizerExtendedEdm.h
@@ -10,11 +10,14 @@
 #include "k4FWCore/DataHandle.h"
 #include "k4Interface/IGeoSvc.h"
 
-// EDM4HEP
+// EDM4HEP & PODIO
 #include "edm4hep/SimTrackerHitCollection.h"
+#include "podio/UserDataCollection.h"
 
 // EDM4HEP extension
 #include "extension/DriftChamberDigiCollection.h"
+#include "extension/DriftChamberDigiLocalCollection.h"
+#include "extension/MCRecoDriftChamberDigiAssociationCollection.h"
 
 // DD4HEP
 #include "DD4hep/Detector.h"  // for dd4hep::VolumeManager
@@ -52,6 +55,10 @@ private:
   DataHandle<edm4hep::SimTrackerHitCollection> m_input_sim_hits{"inputSimHits", Gaudi::DataHandle::Reader, this};
   // Output digitized tracker hit collection name
   DataHandle<extension::DriftChamberDigiCollection> m_output_digi_hits{"outputDigiHits", Gaudi::DataHandle::Writer, this};
+  // Output association between digitized and simulated hit collections
+  DataHandle<extension::MCRecoDriftChamberDigiAssociationCollection> m_output_sim_digi_association{"outputSimDigiAssociation", Gaudi::DataHandle::Writer, this};
+  // Output digitized tracker hit in local coordinates collection name. Only filled in debug mode
+  DataHandle<extension::DriftChamberDigiLocalCollection> m_output_digi_local_hits{"outputDigiLocalHits", Gaudi::DataHandle::Writer, this};
 
   // Detector readout name
   Gaudi::Property<std::string> m_readoutName{this, "readoutName", "CDCHHits", "Name of the detector readout"};
@@ -67,6 +74,13 @@ private:
                                "Spatial resolution in the z direction (from reading out the wires at both sides) [mm]"};
   // xy resolution in mm
   FloatProperty m_xy_resolution{this, "xyResolution", 0.1, "Spatial resolution in the xy direction [mm]"};
+  // Flag to produce debugging distributions
+  Gaudi::Property<bool> m_debugMode{this, "debugMode", false, "Flag to produce debugging distributions"};
+  // Declaration of debugging distributions
+  DataHandle<podio::UserDataCollection<double>> m_leftHitSimHitDeltaDistToWire{"leftHitSimHitDeltaDistToWire", Gaudi::DataHandle::Writer, this}; // mm
+  DataHandle<podio::UserDataCollection<double>> m_leftHitSimHitDeltaLocalZ{"leftHitSimHitDeltaLocalZ", Gaudi::DataHandle::Writer, this}; // mm
+  DataHandle<podio::UserDataCollection<double>> m_rightHitSimHitDeltaDistToWire{"rightHitSimHitDeltaDistToWire", Gaudi::DataHandle::Writer, this}; // mm
+  DataHandle<podio::UserDataCollection<double>> m_rightHitSimHitDeltaLocalZ{"rightHitSimHitDeltaLocalZ", Gaudi::DataHandle::Writer, this}; // mm
 
   // Random Number Service
   IRndmGenSvc* m_randSvc;