+++ DCHdigi/test/runDCHsimpleDigitizerExtendedEdm.py
@@ -7,7 +7,7 @@ podioevent  = FCCDataSvc("EventDataSvc")
 
 from GaudiKernel.SystemOfUnits import MeV, GeV, tesla
 ################## Particle gun setup
-momentum = 5 # in GeV
+momentum = 10 # in GeV
 #thetaMin = 90.25 # degrees
 #thetaMax = 90.25 # degrees
 thetaMin = 20 # degrees
@@ -60,7 +60,7 @@ regiontool = SimG4UserLimitRegion("limits")
 regiontool.volumeNames = ["CDCH"]
 regiontool.OutputLevel = DEBUG
 from GaudiKernel.SystemOfUnits import mm
-regiontool.maxStep = 0.4*mm
+#regiontool.maxStep = 0.4*mm
 
 from Configurables import SimG4UserLimitPhysicsList
 # create overlay on top of FTFP_BERT physics list, attaching fast sim/parametrization process
@@ -122,20 +122,35 @@ from Configurables import DCHsimpleDigitizerExtendedEdm
 dch_digitizer = DCHsimpleDigitizerExtendedEdm("DCHsimpleDigitizerExtendedEdm",
     inputSimHits = savetrackertool.SimTrackHits.Path,
     outputDigiHits = savetrackertool.SimTrackHits.Path.replace("sim", "digi"),
+    outputSimDigiAssociation = "DC_simDigiAssociation",
     readoutName = "CDCHHits",
     xyResolution = 0.1, # mm
     zResolution = 1, # mm
-    OutputLevel=INFO
+    debugMode = False,
+    OutputLevel = INFO
 )
+
+# Derive performance quantities
+#from Configurables import DCHsimpleDigitizerExtendedEdmPerformance
+#dch_perf = DCHsimpleDigitizerExtendedEdmPerformance("DCHsimpleDigitizerExtendedEdmPerformance",
+#        inputSimHits = savetrackertool.SimTrackHits.Path,
+#        inputDigiHits = dch_digitizer.outputDigiHits,
+#        inputSimDigiAssociation = dch_digitizer.outputSimDigiAssociation
+#)
+
+
 
 ################ Output
 from Configurables import PodioOutput
 out = PodioOutput("out",
                   OutputLevel=INFO)
 out.outputCommands = ["keep *"]
+if not dch_digitizer.debugMode:
+    out.outputCommands.append("drop *HitSimHitDelta*")
+    out.outputCommands.append("drop outputDigiLocalHits")
 
 import uuid
-out.filename = "output_simplifiedDriftChamber_MagneticField_"+str(magneticField)+"_pMin_"+str(momentum*1000)+"_MeV"+"_ThetaMinMax_"+str(thetaMin)+"_"+str(thetaMax)+"_pdgId_"+str(pdgCode)+"_stepLength_"+str(regiontool.maxStep)+".root"
+out.filename = "output_simplifiedDriftChamber_MagneticField_"+str(magneticField)+"_pMin_"+str(momentum*1000)+"_MeV"+"_ThetaMinMax_"+str(thetaMin)+"_"+str(thetaMax)+"_pdgId_"+str(pdgCode)+"_stepLength_default.root"
 
 #CPU information
 from Configurables import AuditorSvc, ChronoAuditor
@@ -159,10 +174,11 @@ ApplicationMgr(
               hepmc_converter,
               geantsim,
               dch_digitizer,
+              #dch_perf,
               out
               ],
     EvtSel = 'NONE',
-    EvtMax   = 10,
+    EvtMax   = 100,
     ExtSvc = [geoservice, podioevent, geantservice, audsvc],
     StopOnSignal = True,
  )