+++ doc/ReleaseNotes.md
@@ -1,3 +1,8 @@
+# v00-02-00
+
+* 2024-02-12 tmadlener ([PR#16](https://github.com/key4hep/k4RecTracker/pull/16))
+  - Add a schema_version to the data model extension to keep it working after [podio#556](https://github.com/AIDASoft/podio/pull/556)
+
 # v00-01
 
 * This file is also automatically populated by the tagging script